% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pnam_wrapper_functions.R
\name{coef.pnam}
\alias{coef.pnam}
\title{Extract the ML parameter estimates from Panel Network Autocorrelation Model Fits}
\usage{
\method{coef}{pnam}(object, ...)
}
\arguments{
\item{object}{An object of class "pnam".}

\item{...}{Additional arguments for other methods.}
}
\description{
This function extracts the Maximum Likelihood (ML) parameter estimates from estimated
panel network autocorrelation Model fits.
}
