% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pnam_wrapper_functions.R
\name{fitted.pnam}
\alias{fitted.pnam}
\title{Fitted values from \code{pnam} model fits}
\usage{
\method{fitted}{pnam}(object, ...)
}
\arguments{
\item{object}{An object of class "pnam".}

\item{...}{Additional arguments for other methods.}
}
\description{
This function returns the model fitted values for a \code{pnam} model fit.
}
