% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pnam_wrapper_functions.R
\name{logLik.dream_rem}
\alias{logLik.dream_rem}
\title{Extract the model log-likelihood from  `pnam` Fits}
\usage{
\method{logLik}{dream_rem}(object, ..., REML = FALSE)
}
\arguments{
\item{object}{An object of class "pnam".}

\item{...}{Additional arguments for other methods.}

\item{REML}{From the generic `logLik` function. Set to FALSE and does not
need to changed by the user.}
}
\description{
This function extracts the model loglikelhood from estimated
panel netowrk autocorrelation model fits.
}
