% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netpanel_wrapper_functions.R
\name{logLik.pnam}
\alias{logLik.pnam}
\title{Extract the estimated model log-likelihood from \code{pnam} model fits}
\usage{
\method{logLik}{pnam}(object, ..., REML = FALSE)
}
\arguments{
\item{object}{An object of class \code{pnam}.}

\item{...}{Additional arguments for other methods.}

\item{REML}{From the generic \code{logLik} function. Set to FALSE and does not
need to changed by the user.}
}
\value{
The estimated model log-likelihood.
}
\description{
This function extracts the estimated model loglikelhood from a \code{pnam} model fit.
}
\examples{
data("net1", package = "netpanel") #the first panel network
data("simulated.data", package = "netpanel") #the simulated variables

# a fixed effects panel network autocorrelation model
fe.pnam <- mlpnam(Y~x1+x2+x3, net.formula = ~ net1,
               data =  simulated.data, time = ~panel,
               actor = ~unit, model = "fixed",
               fixed.effect = "actor")
logLik(fe.pnam) #the estimated model log-likelihood.
}
