% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pnam_wrapper_functions.R
\name{mlpnam}
\alias{mlpnam}
\title{Fit a Maximum Likelihood Panel Network Autocorrelation Model (PNAM)}
\usage{
mlpnam(
  reg.formula,
  net.formula,
  data,
  time,
  actor,
  model = c("fixed", "random", "dynamic"),
  fixed.effect = c("actor", "time", "two-way"),
  dynamic.lag = c("network", "outcome", "network x outcome"),
  errors = c("idiosyncratic", "autocorrelated"),
  autocorrelated.network = NULL,
  optim.method = "L-BFGS-B",
  optim.control = list(),
  ...
)
}
\arguments{
\item{reg.formula}{A \code{formula} object that is the symbolic description of the network autocorrelation model to be
fitted. The dependent and exogenous regressors are extracted from the \code{data} argument.
This is the same argument found in \code{\link[stats]{lm}} and \code{\link[stats]{glm}}.}

\item{net.formula}{A \code{formula} object that is the symbolic description of the network structure for the network
autocorrelation model, which allows multiple networks to be included. The dependent variable should not be defined. Each of the included covariates
should be a list of length t (i.e., the number of time points/panels) and each element
should be an N x N matrix, where N is the number of actors/units in the network. The included networks
should be row-normalized. The networks are extracted from the environment.}

\item{data}{An object of class \code{data.frame} that contains the variables from the \code{reg.formula},
\code{time}, and \code{actor} arguments. The \code{data.frame} should be stacked by time, and ordered within
time with respect to their position in the network adjacency matrices.}

\item{time}{A \code{formula} object where the variable that represents the time index of the observation is on the
right hand side of ~. The variable is extracted from the \code{data} argument.}

\item{actor}{A \code{formula} object where the variable that represents the actor/unit index of the observation is on the
right hand side of ~. The variable is extracted from the \code{data} argument.}

\item{model}{The type of panel network autocorrelation model to estimate (see the Details section). "fixed" indicates
that a within-estimator fixed effects model should be estimated. The demeaning of the data structure is based upon the
\code{fixed.effect} argument. "random" indicates that a random-intercept model should be estimated with random effects
with respect to the actors/units. "dynamic" indicates that a lagged model will be estimated, and the specific lag
structure is based upon the \code{dynamic.lag} argument.}

\item{fixed.effect}{\strong{optional}. When \code{model} is set to "fixed", how should the data structure be demeaned? "actor" indicates
the data will be demeaned by the unit-specific average. "time" indicates that the data will be time demeaned (i.e.,
the i value at time t will be subtracted by the mean value at time t across units). "two-way" indicates that
two-way fixed effects will be included (i.e., time and unit fixed effects).}

\item{dynamic.lag}{\strong{optional}. When \code{model} is set to "dynamic", what lag structure should be included? The option
"network" will add the time lag for each included network structure in the \code{net.formula} argument. The option
"outcome" will add the time lag for the outcome variable (Y) based on the \code{reg.formula} argument. The "network x outcome"
option will add the time lags for the outcome and network variables.}

\item{errors}{What is the time-varying residual structure? "idiosyncratic" indicates that the time-varying residuals are assumed to
be drawn from the following distribution: N(0,sigma2), where sigma2 is estimated variance. "autocorrelated" assumes that the
time-varying residuals exhibit network autocorrelation and the model become a network lag x error model. See the Details section.}

\item{autocorrelated.network}{\strong{optional}. When \code{errors} is set to "autocorrelated", a \code{list} object with length t, where
each element is an N x N matrix and represents the network to be used in the network error model. See the Details section.}

\item{optim.method}{\strong{optional}. The numeric optimization procedure to be employed by \code{\link[stats]{optim}}. Defaults to
"L-BFGS-B" and, in general, should not be changed.}

\item{optim.control}{\strong{optional}. A list control options for the \code{\link[stats]{optim}} function. Please see the
\code{\link[stats]{optim}} function for the set of options.}

\item{...}{Additional arguments.}
}
\value{
An object of class "dream_rem" as a list containing the following components:
\itemize{
\item \code{N} - The number of unique units in the panel network dataset.
\item \code{t} - The number of panels (time periods) in the panel network dataset.
\item \code{type} - The type of model estimate (e.g., random effects).
\item \code{error.structure} - The assumed structure for the time-varying errors.
\item \code{fe.type} - If a fixed effects model is estimated, the type of fixed effects demeaning.
\item \code{lag.type} - If a dynamic model is estimated, the type of lag structure included in the model.
\item \code{max.eigen} - The maximum absolute eigenvalue of the rho*W matrix.
\item \code{lambda} - If the time-varying errors are assumed to be autocorrelated, the estimated value for lambda.
\item \code{rho} - The vector of estimated maximum likelihood rho parameters.
\item \code{vcov.rho} - The asymptotic variance-covariance of the estimated rho parameters.
\item \code{coefficients} - The vector of estimated exogenous ML parameters.
\item \code{vcov} - The asymptotic variance-covariance of the exogenous ML parameters.
\item \code{sigma2} - The estimated variance of the time-varying residual.
\item \code{residuals} - The residuals of the fitted model.
\item \code{fitted.values} - The fitted values from the estimated model.
\item \code{logLik} - The estimated model log-likelihood.
\item \code{convergence} - The converge value returned from the optim function. 0 indicates successful convergence.
\item \code{netfilter} - The filtering matrix of the estimated model \eqn{I_{NT} - \sum_{j} rho_j*A_j }, where \eqn{A_j} is the jth network from the
\code{net.formula} argument.
\item \code{k} - The total number of estimated parameters.
\item \code{df} - The model degrees of freedom.
\item \code{AIC} - The AIC of the estimated PNAM.
\item \code{BIC} - The BIC of the estimated PNAM.
\item \code{response} - The outcome vector (\eqn{Y}).
\item \code{covariates} - The covariate matrix (\eqn{X}).
\item \code{optim.information} - The returned list from the optim function.
\item \code{error.dis} - If a random effects model is estimated, a list that contains the
residual variances for the time-varying residuals (idiosyncratic), the unit-/actor-specific
residual variances (unit), and the estimated ratio of the variances (theta).
\item \code{theta} - If a random effects model is estimated, the estimated value for
theta (the ratio of the unit and idiosyncratic variances.)
\item \code{vcov.theta} - If a random effects model is estimated, the asymptotic variance of
the estimated theta value.
}
}
\description{
A function to estimate a Maximum Likelihood Panel Network Autocorrelation Model (MLPNAM). The function
allows for users to input one or multiple networks to represent the network structure. If multiple networks
are included the PNAM becomes an M-STAR model (see XXXX). In addition, the function allows for (1) fixed and
random effects and (2) the inclusion of lagged outcome and network terms. Finally, the function
allows for the time-varying errors to exhibit network autocorrelation, that is, a network lag and error model.
}
\details{
need to write this.......
}
\examples{
data("network.1", package = "netpanel") #the first panel network
data("network.2", package = "netpanel") #the second panel network
data("error.network", package = "netpanel") #the third panel network
data("simulated.data", package = "netpanel")

# a random effects panel network autocorrelation model
re.pnam <- mlpnam(Y~x1+x2+x3, net.formula = ~ net1 + net2,
               data =  simulated.data, time = ~panel,
               actor = ~unit, model = "random")
summary(re.pnam)              

# a fixed effects panel network autocorrelation model with autocorrelated errors
fe.pnam.error <- mlpnam(Y~x1+x2+x3, net.formula = ~  net1,
               data =  simulated.data, time = ~panel,
               actor = ~unit, model = "fixed",
               fixed.effect = "actor", errors = "autocorrelated",
               autocorrelated.network = error.net)
summary(fe.pnam.error)   

# a dynamic panel network autocorrelation model with autocorrelated errors
dyn.pnam <- mlpnam(Y~x1+x2+x3, net.formula = ~  net1 + net2,
               data =  simulated.data, time = ~panel,
               actor = ~unit, model = "dynamic",
               dynamic.lag = "outcome")
summary(dyn.pnam)   


}
\references{
add to this here....
}
\seealso{
\code{\link{coef.pnam}}, \code{\link{vcov.pnam}}, \code{\link{logLik.pnam}},
\code{\link{AIC}}, \code{\link{residuals.pnam}}, \code{\link{fitted.pnam}},
\code{\link{netimpacts}}.
}
