% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pnam_data.R
\docType{data}
\name{net1}
\alias{net1}
\title{Simulated panel networks (1)}
\format{
\subsection{\code{network.1}}{

A \code{list} of length 10 that contains simulated row-normalized 50 x 50 social
adjacency matrices for each time point.
}
}
\source{
Butts CT (2024). \emph{sna: Tools for Social Network Analysis}. doi:10.32614/CRAN.package.sna
\url{https://doi.org/10.32614/CRAN.package.sna}, R package version 2.8, \url{https://CRAN.R-project.org/package=sna}.
}
\usage{
data(network.1)
}
\description{
A \code{list} of length 10 that contains simulated row-normalized 50 x 50 social
adjacency matrices for each time point. The networks are generated with the
\code{\link[sna]{rgraph}} function in the sna package (Butts 2024).
}
\keyword{datasets}
