% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netpanel_data.R
\docType{data}
\name{net2}
\alias{net2}
\title{A \code{list} of simulated panel social networks (2)}
\format{
\subsection{\code{net2}}{

A \code{list} of length 10 where each element is an simulated row-normalized 50 x 50 social
adjacency matrix, where a non-zero \eqn{W_{ijt}} entry denotes a directed relationship
from \eqn{i} to \eqn{j} at time \eqn{t}.
}
}
\usage{
data(net2)
}
\description{
A \code{list} of length 10 where each element is an simulated row-normalized 50 x 50 social
adjacency matrix, where a non-zero \eqn{W_{ijt}} entry denotes a directed relationship
from \eqn{i} to \eqn{j} at time \eqn{t}.
}
\keyword{datasets}
