% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netpanel_wrapper_functions.R
\name{netimpacts}
\alias{netimpacts}
\title{Compute the direct, indirect, and total impacts from a fitted \code{pnam} object}
\usage{
netimpacts(object, return.impact.matrix = FALSE, digits = 5)
}
\arguments{
\item{object}{An object of class \code{pnam}.}

\item{return.impact.matrix}{TRUE/FALSE. If TRUE, the function will also return the estimated
impact matrix for each estimated covariate. Set to FALSE by default.}

\item{digits}{The number of digits to round the estimates after the decimal point.}
}
\value{
A \code{data.frame} object that stores the direct, indirect, and total impact for each estimated effect in the fitted \code{pnam} model.
}
\description{
This function returns the estimated average direct, indirect, and total
impacts from a fitted panel network autocorrelation model.
}
\details{
In classical linear regression, the marginal effect of a regressor, \eqn{x_1}, on the
outcome, \eqn{Y}, is \eqn{\hat{\beta_1}}. In panel network (and spatial) effects models, the
marginal effect is:
\deqn{\frac{\partial Y}{\partial x_1} = (I_{NT} - \sum_k \hat{\rho_k}W^{k})^{-1}\hat{\beta_1}}
The direct impact is the average of the diagonal value of the above matrix, the
indirect impact is the average of the row sum of the off-diagonal values, and
the total impact is sum of both. The function will also return the impact matrix for each
estimated effect if \code{return.impact.matrix} is set to TRUE.
}
\examples{
data("net1", package = "netpanel") #the first panel network
data("simulated.data", package = "netpanel") #the simulated variables

# a fixed effects panel network autocorrelation model
fe.pnam <- mlpnam(Y~x1+x2+x3, net.formula = ~ net1,
               data =  simulated.data, time = ~panel,
               actor = ~unit, model = "fixed",
               fixed.effect = "actor")
netimpacts(fe.pnam) #compute the impact values

}
