% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netpanel-package.R
\docType{package}
\name{netpanel-package}
\alias{netpanel-package}
\title{netpanel: Maximum Likelihood Estimation Routines for Panel Network Autocorrelation Models}
\description{
This package contains maximum likelihood estimation routines for panel network autocorrelation models. The package contains functionality for fixed and random effects panel network autocorrelation model, in addition to dynamic panel network autocorrelation that contains time lags of the response and networks. Moreover, the package allows for multiple networks to be fitted. For information on the types of models estimated in this package, please see Anselin (1988) \doi{10.1007/978-94-015-7799-1}, Cook et al. (2023) \doi{10.1017/S0003055422000272}, Hays et al. (2010) \doi{10.1016/j.stamet.2009.11.005}, Lee and Yu (2012) \doi{10.1111/j.1468-2354.2012.00724.x}, Millo (2014) \doi{10.1016/j.csda.2013.07.024}, Millo and Piras (2012) \doi{10.18637/jss.v047.i01}, and Wang and Yu (2015) \doi{10.1016/j.econlet.2015.01.021}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/kevinCarson/netpanel}
  \item Report bugs at \url{https://github.com/kevinCarson/netpanel/issues}
}

}
\author{
\strong{Maintainer}: Kevin Carson \email{kacarson@arizona.edu}

}
\keyword{internal}
