% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pnam_estimation_routines.R
\name{panelnam.dynam.error}
\alias{panelnam.dynam.error}
\title{A Function To Find the ML Estimates for a Dynamic Panel Network Autocorrelation Model with Network Error Autocorrelation}
\usage{
panelnam.dynam.error(
  X,
  Y,
  W,
  n,
  t,
  dyanmic.lag = c("network", "outcome", "network x outcome"),
  optim.method,
  optim.control,
  W2
)
}
\arguments{
\item{X}{the co variate NT x K matrix.}

\item{Y}{The NT x 1 vector of outcomes.}

\item{W}{A list of J block-diagonal NT x NT matrix}

\item{n}{the number of units}

\item{t}{the number of time points}

\item{dyanmic.lag}{the type of lags to add to the model.}

\item{optim.method}{The user-provided optim search method}

\item{optim.control}{The list of user-provided optim control parameters}

\item{W2}{the NT x NT block-diagonal matrix for the error network autocorrelation.}
}
\value{
A list that store the maximum likelihood pnam estimates.
}
\description{
The dynamic panel network autocorrelation model conditions on the first observations (see Hays et al. 2010).
}
\keyword{internal}
