% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pnam_estimation_functions.R
\name{panelnam.dynam.wae}
\alias{panelnam.dynam.wae}
\title{Estimate a Dynamic Panel Network Autocorrelation Model with Autocorrelated Errors}
\usage{
panelnam.dynam.wae(X, Y, W, B, n, t)
}
\arguments{
\item{X}{the covariate NT x K matrix.}

\item{Y}{The vector of outcomes.}

\item{W}{The block-diagonal NT x NT matrix}

\item{B}{The block-diagonal NT x NT matrix for the spatial error term}

\item{n}{the number of units}

\item{t}{the number of time points}
}
\value{
The maximum likelihood estimates.
}
\description{
Estimate a Dynamic Panel Network Autocorrelation Model with Autocorrelated Errors
}
