% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pnam_estimation_routines.R
\name{panelnam.fixed.error}
\alias{panelnam.fixed.error}
\title{A Function To Find the ML Estimates for a Fixed Effects Panel Network Autocorrelation Model with Network Error Autocorrelation}
\usage{
panelnam.fixed.error(
  X,
  Y,
  W,
  n,
  t,
  id.index,
  time.index,
  fe,
  optim.method,
  optim.control,
  W2
)
}
\arguments{
\item{X}{the co variate NT x K matrix.}

\item{Y}{The NT x 1 vector of outcomes.}

\item{W}{A list of J block-diagonal NT x NT matrix}

\item{n}{the number of units}

\item{t}{the number of time points}

\item{id.index}{the vector of indices to map the unique units.}

\item{time.index}{the vector of indices to map the time units.}

\item{fe}{the type of demeaning for the fixed effects}

\item{optim.method}{The user-provided optim search method}

\item{optim.control}{The list of user-provided optim control parameters}

\item{W2}{the NT x NT block-diagonal matrix for the error network autocorrelation.}
}
\value{
A list that store the maximum likelihood pnam estimates.
}
\description{
A Function To Find the ML Estimates for a Fixed Effects Panel Network Autocorrelation Model with Network Error Autocorrelation
}
\keyword{internal}
