% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pnam_estimation_routines.R
\name{panelnam.random.error}
\alias{panelnam.random.error}
\title{A Function To Find the ML Estimates for a Panel Network Autocorrelation Model with Network Error Autocorrelation and Unit-Level Random Effects}
\usage{
panelnam.random.error(X, Y, W, n, t, optim.control, optim.method, W2)
}
\arguments{
\item{X}{the co variate NT x K matrix.}

\item{Y}{The NT x 1 vector of outcomes.}

\item{W}{A list of J block-diagonal NT x NT matrix}

\item{n}{the number of units}

\item{t}{the number of time points}

\item{optim.control}{The list of user-provided optim control parameters}

\item{optim.method}{The user-provided optim search method}

\item{W2}{the NT x NT block-diagonal matrix for the error network autocorrelation.}

\item{dyanmic.lag}{the type of lags to add to the model.}
}
\value{
A list that store the maximum likelihood pnam estimates.
}
\description{
A Function To Find the ML Estimates for a Panel Network Autocorrelation Model with Network Error Autocorrelation and Unit-Level Random Effects
}
\keyword{internal}
