% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pnam_estimation_routines.R
\name{panelnam.random}
\alias{panelnam.random}
\title{A Function To Find the ML Estimates for a Panel Network Autocorrelation Model with Unit-Level Random Effects}
\usage{
panelnam.random(X, Y, W, n, t, optim.control, optim.method)
}
\arguments{
\item{X}{the co variate NT x K matrix.}

\item{Y}{The NT x 1 vector of outcomes.}

\item{W}{A list of J block-diagonal NT x NT matrix}

\item{n}{the number of units}

\item{t}{the number of time points}

\item{optim.control}{The list of user-provided optim control parameters}

\item{optim.method}{The user-provided optim search method}
}
\value{
A list that store the maximum likelihood pnam estimates.
}
\description{
A Function To Find the ML Estimates for a Panel Network Autocorrelation Model with Unit-Level Random Effects
}
\keyword{internal}
