% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pnam_estimation_functions.R
\name{panelnamrule}
\alias{panelnamrule}
\title{Estimate a Panel Network Autocorrelation Model with Unit-Level Random Effects}
\usage{
panelnamrule(X, Y, W, n, t)
}
\arguments{
\item{X}{the covariate NT x K matrix.}

\item{Y}{The vector of outcomes.}

\item{W}{The block-diagonal NT x NT matrix}

\item{n}{the number of units}

\item{t}{the number of time points}
}
\value{
The maximum likelihood estimates.
}
\description{
Estimate a Panel Network Autocorrelation Model with Unit-Level Random Effects
}
