% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pnam_estimation_functions.R
\name{panelnamrulewae}
\alias{panelnamrulewae}
\title{Estimate a Panel Network Autocorrelation Model with Unit-Level Random Effects and Autocorrelated Errors}
\usage{
panelnamrulewae(X, Y, W, W1, n, t)
}
\arguments{
\item{X}{the covariate NT x K matrix.}

\item{Y}{The vector of outcomes.}

\item{W}{The block-diagonal NT x NT matrix}

\item{W1}{The block-diagonal NT x NT matrix for the spatial error term}

\item{n}{the number of units}

\item{t}{the number of time points}
}
\value{
The maximum likelihood estimates.
}
\description{
Estimate a Panel Network Autocorrelation Model with Unit-Level Random Effects and Autocorrelated Errors
}
