% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pnam_wrapper_functions.R
\name{residuals.pnam}
\alias{residuals.pnam}
\title{Residuals for \code{pnam} model fits}
\usage{
\method{residuals}{pnam}(object, ...)
}
\arguments{
\item{object}{An object of class "pnam".}

\item{...}{Additional arguments for other methods.}
}
\description{
This function returns the model residuals for a \code{pnam} model fit.
}
\examples{
data("network.1", package = "netpanel") #the first panel network
data("simulated.data", package = "netpanel")

# a fixed effects panel network autocorrelation model
fe.pnam <- mlpnam(Y~x1+x2+x3, net.formula = ~ net1,
               data =  simulated.data, time = ~panel,
               actor = ~unit, model = "fixed",
               fixed.effect = "actor")
residuals(fe.pnam)


}
