% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pnam_data.R
\docType{data}
\name{simulated.data}
\alias{simulated.data}
\title{Simulated panel network datasets}
\format{
\subsection{\code{simulated.data}}{

A simulated panel network dataset that contains the outcome (Y) and three exogenous
regressors (x1, x3, and x3), where x1 and x3 are drawn from the standard normal
distribution, and x2 is a random bernoulli (binary) regressor.
}
}
\usage{
data(simulated.data)
}
\description{
A simulated panel network dataset that contains the outcome (Y) and three exogenous
regressors (x1, x3, and x3), where x1 and x3 are drawn from the standard normal
distribution, and x2 is a random bernoulli (binary) regressor. The \code{panel}
variable represents the time of data collection and \code{unit} indexes the
actor.
}
\keyword{datasets}
