% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pnam_wrapper_functions.R
\name{summary.pnam}
\alias{summary.pnam}
\title{Summary Method for \code{pnam} Objects}
\usage{
\method{summary}{pnam}(object, digits = 5, ...)
}
\arguments{
\item{object}{An object of class "pnam".}

\item{digits}{The number of digits to print after the decimal point.}

\item{...}{Additional arguments (currently unused).}
}
\value{
A list of summary statistics for the panel network autocorrelation model, such
as the estimated log-likelihood and the estimated parameters.
}
\description{
Summarizes the results of a panel network autocorrelation model.
}
