% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netpanel_wrapper_functions.R
\name{vcov.pnam}
\alias{vcov.pnam}
\title{Extract the asymptotic variance-covariance matrix from \code{pnam} model fits}
\usage{
\method{vcov}{pnam}(object, ...)
}
\arguments{
\item{object}{An object of class \code{pnam}.}

\item{...}{Additional arguments for other methods.}
}
\value{
The asymptotic variance-covariance matrix for the exogenous covariates.
}
\description{
This function extracts the asymptotic variance-covariance matrix from
\code{pnam} model fits for the exogenous covariates. For the endogenous
variance matrices (such as the variance-covariance matrix for \eqn{\rho}), please extract those
from the model fit. For example, the variance matrix for \eqn{\rho}
can be extracted as \code{model$vcov.rho}, where \code{model} is the \code{pnam} object name.
}
\examples{
data("net1", package = "netpanel") #the first panel network
data("simulated.data", package = "netpanel") #the simulated variables

# a fixed effects panel network autocorrelation model
fe.pnam <- mlpnam(Y~x1+x2+x3, net.formula = ~ net1,
               data =  simulated.data, time = ~panel,
               actor = ~unit, model = "fixed",
               fixed.effect = "actor")
vcov(fe.pnam) #the asymptotic variance-covariance matrix
}
