% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netpanel_wrapper_functions.R
\name{coef.pnam}
\alias{coef.pnam}
\title{Extract the ML parameter estimates from \code{pnam} model fits}
\usage{
\method{coef}{pnam}(object, ...)
}
\arguments{
\item{object}{An object of class "pnam".}

\item{...}{Additional arguments for other methods.}
}
\value{
The vector of estimated parameter values for the exogenous covariates.
}
\description{
This function extracts the Maximum Likelihood (ML) parameter estimates for the
exogenous covariates from a \code{pnam} model fit.
}
\examples{
data("net1", package = "netpanel") #the first panel network
data("simulated.data", package = "netpanel") #the simulated variables
#' a fixed effects panel network autocorrelation model
fe.pnam <- mlpnam(Y~x1+x2+x3, net.formula = ~ net1,
               data =  simulated.data, time = ~panel,
               actor = ~unit, model = "fixed",
               fixed.effect = "actor")
coef(fe.pnam) #the mle parameter values
}
