% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netpanel_wrapper_functions.R
\name{fitted.pnam}
\alias{fitted.pnam}
\title{Fitted values from \code{pnam} model fits}
\usage{
\method{fitted}{pnam}(object, ...)
}
\arguments{
\item{object}{An object of class \code{pnam}.}

\item{...}{Additional arguments for other methods.}
}
\value{
A \code{data.frame} object that stores the fitted values from a \code{pnam} model fit, where the column "ols.fitted" stores the
linear regression fitted values and "network.fitted" stores the network regression fitted values (see above).
}
\description{
This function returns the model fitted values, \eqn{\hat{y}}, for a fitted \code{pnam} model. The function returns
two types of fitted values to the user: linear regression fitted values and network regression fitted values. The
network regression fitted values are formulated as: \eqn{\hat{y} = (I_{NT} - \sum_i \hat{\rho}_i W^i)^{-1}(X\hat{\beta})}, where \eqn{W^i}
is the \eqn{i^{th}} included network. In comparison, the linear regression fitted values are formulated as:
\eqn{\hat{y} = X\hat{\beta}}.
}
\examples{
data("net1", package = "netpanel") #the first panel network
data("simulated.data", package = "netpanel") #the simulated variables

# a fixed effects panel network autocorrelation model
fe.pnam <- mlpnam(Y~x1+x2+x3, net.formula = ~ net1,
               data =  simulated.data, time = ~panel,
               actor = ~unit, model = "fixed",
               fixed.effect = "actor")
fitted(fe.pnam) #the data.frame of fitted values

}
