% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netpanel_wrapper_functions.R
\name{mlpnam}
\alias{mlpnam}
\title{Fit a Maximum Likelihood Panel Network Autocorrelation Model (PNAM)}
\usage{
mlpnam(
  reg.formula,
  net.formula,
  data,
  time,
  actor,
  model = c("fixed", "random", "dynamic"),
  fixed.effect = c("actor", "time", "two-way"),
  dynamic.lag = c("network", "outcome", "network x outcome"),
  errors = c("idiosyncratic", "autocorrelated"),
  autocorrelated.network = NULL,
  optim.method = "L-BFGS-B",
  optim.control = list(),
  ...
)
}
\arguments{
\item{reg.formula}{A \code{formula} object that is the symbolic description of the network autocorrelation model to be
fitted. The dependent and exogenous regressors are extracted from the \code{data} argument.
This is the same argument found in \code{\link[stats]{lm}} and \code{\link[stats]{glm}}.}

\item{net.formula}{A \code{formula} object that is the symbolic description of the network structure for the network
autocorrelation model, which allows multiple networks to be included. The dependent variable should not be defined. Each of the included covariates
should be a list of length t (i.e., the number of time points/panels) and each element
should be an N x N matrix, where N is the number of actors/units in the network. The included networks
should be row-normalized. The networks are extracted from the environment.}

\item{data}{An object of class \code{data.frame} that contains the variables from the \code{reg.formula},
\code{time}, and \code{actor} arguments. The \code{data.frame} should be stacked by time, and ordered within
time with respect to their position in the network adjacency matrices. If you are unsure how the
stacking of observations should be constructed, please see the structure of the
\code{\link{simulated.data}} \code{netpanel} data object.}

\item{time}{A \code{formula} object where the variable that represents the time index of the observation is on the
right hand side of ~. The variable is extracted from the \code{data} argument.}

\item{actor}{A \code{formula} object where the variable that represents the actor/unit index of the observation is on the
right hand side of ~. The variable is extracted from the \code{data} argument.}

\item{model}{The type of panel network autocorrelation model to be estimated (see the Details section). "fixed" indicates
that a within-estimator fixed effects model should be estimated. The demeaning of the data structure is based upon the
\code{fixed.effect} argument. "random" indicates that a random-intercept model should be estimated with random effects
with respect to the actors/units. "dynamic" indicates that a lagged model will be estimated, and the specific lag
structure is based upon the \code{dynamic.lag} argument.}

\item{fixed.effect}{\strong{optional}. When \code{model} is set to "fixed", how should the data structure be demeaned? "actor" indicates
the data will be demeaned by the unit-specific average. "time" indicates that the data will be time demeaned (i.e.,
the \eqn{x_{i1t}} value at time \eqn{t} will be subtracted by the mean of \eqn{x_{1t}}, which is the vector of
\eqn{x_1} values for all units at time \eqn{t}). "two-way" indicates that
two-way fixed effects will be included (i.e., time and unit fixed effects).}

\item{dynamic.lag}{\strong{optional}. When \code{model} is set to "dynamic", what lag structure should be included? The option
"network" will add the time lag for each included network structure in the \code{net.formula} argument. The option
"outcome" will add the time lag for the outcome variable (Y) based on the \code{reg.formula} argument. The "network x outcome"
option will add the time lags for the outcome and network variables.}

\item{errors}{What is the assumed structure of time-varying errors? "idiosyncratic" indicates that the time-varying errors are assumed to
be drawn from the following distribution: \eqn{N(0,\sigma^2)}, where \eqn{\sigma^2} is the variance. "autocorrelated" assumes that the
time-varying errors exhibit network autocorrelation and the model will become a network mixed effects model. See the Details section.}

\item{autocorrelated.network}{\strong{optional}. When \code{errors} is set to "autocorrelated", a \code{list} object with length t, where
each element is an N x N matrix and represents the network to be used in the network error model. See the Details section.}

\item{optim.method}{\strong{optional}. The numeric optimization procedure to be employed by \code{\link[stats]{optim}}. Defaults to
"L-BFGS-B" and, in general, should not be changed.}

\item{optim.control}{\strong{optional}. A list control options for the \code{\link[stats]{optim}} function. Please see the
\code{\link[stats]{optim}} function for the set of options.}

\item{...}{Additional arguments.}
}
\value{
An object of class \code{pnam} as a \code{list} containing the following elements:
\itemize{
\item \code{N} - The number of unique units in the panel network dataset.
\item \code{t} - The number of panels (time periods) in the panel network dataset.
\item \code{type} - The type of model estimated per the \code{model} argument.
\item \code{error.structure} - The assumed structure for the time-varying errors per the \code{errors} argument.
\item \code{fe.type} - If a fixed effects model is estimated, the type of fixed effects demeaning per the \code{fixed.effect} argument.
\item \code{lag.type} - If a dynamic model is estimated, the type of lag structure included in the model per the \code{dynamic.lag} argument
\item \code{max.eigen} - The maximum absolute eigenvalue of the rho*W matrix.
\item \code{lambda} - If the time-varying errors are assumed to be autocorrelated, the estimated value for lambda.
\item \code{rho} - The vector of estimated maximum likelihood rho parameters.
\item \code{vcov.rho} - The asymptotic variance-covariance of the estimated rho parameters.
\item \code{coefficients} - The vector of estimated exogenous ML parameters.
\item \code{vcov} - The asymptotic variance-covariance of the exogenous ML parameters.
\item \code{sigma2} - The estimated variance of the time-varying residual.
\item \code{residuals} - The residuals of the fitted model.
\item \code{fitted.values} - The fitted values from the estimated model.
\item \code{logLik} - The estimated model log-likelihood.
\item \code{convergence} - The converge value returned from the optim function. 0 indicates successful convergence.
\item \code{netfilter} - The filtering matrix of the estimated model (\eqn{I_{NT} - \{\sum_{j} rho_j*W^j \}}), where \eqn{W^j} is the jth network from the
\code{net.formula} argument.
\item \code{k} - The total number of estimated parameters.
\item \code{df} - The model degrees of freedom.
\item \code{AIC} - The AIC of the estimated PNAM.
\item \code{BIC} - The BIC of the estimated PNAM.
\item \code{response} - The outcome vector (\eqn{Y}).
\item \code{covariates} - The covariate matrix (\eqn{X}).
\item \code{optim.information} - The returned list from the optim function.
\item \code{error.dis} - If a random effects model is estimated, a list that contains the
residual variances for the time-varying residuals (idiosyncratic), the unit-/actor-specific
residual variances (unit), and the estimated ratio of the variances (theta).
\item \code{theta} - If a random effects model is estimated, the estimated value for
theta (the ratio of the unit and idiosyncratic variances.)
\item \code{vcov.theta} - If a random effects model is estimated, the asymptotic variance of
the estimated theta value.
}
}
\description{
A function that contains Maximum Likelihood estimation routines to fit panel network autocorrelation models (MLPNAM).
The function allows for (1) fixed and random actor-level effects, (2) the inclusion of time lagged outcome and network terms,
(3) multiple networks to construct the network (spatial) filtering matrix, and (4) for the time-varying
errors to exhibit network autocorrelation, that is, for the model to become a network mixed effects model (i.e., error + lag model). Please
see the Details section for more information.
}
\details{
This function fits a panel network autocorrelation model with time-varying social adjacency matrices
via Maximum Likelihood Estimation. The estimated model takes the form:
\deqn{Y_t = \rho_1 W^{t}_{1}Y_t + \cdots + \rho_k W^{t}_{k}Y_t + X_t\beta + v_t, \ \ t = 1,2,\cdots,T }
\deqn{v_t = \mu_N + e_t,\ \ t = 1,2,\cdots,T   }
where \eqn{Y_t} is the outcome vector for \eqn{N} actors at time \eqn{t}, \eqn{\rho_k W^{t}_{k}Y_t} is the
network autocorrelation term for the \eqn{k^{th}} social network at time \eqn{t} (taken from the \code{net.formula}
argument), \eqn{X_t} are the set of exogenous time-varying (and invariant) regressors (taken from the \code{reg.formula}
argument). Additionally, \eqn{v_t} is the time-varying error term that is the sum
of the time-invariant unit-specific unobserved factors, \eqn{\mu_N}, thought to explain \eqn{Y_t} and \eqn{e_t} are the
time- and unit-varying unobserved factors. Dependent upon the \code{model} argument, the function will (1) estimate a fixed-effects
network regression, where the within estimator demeans the data by time, actor, or both, (2) estimate a random effects
network regression, or (3) estimate a dynamic panel network autocorrelation where the time lags of the outcome
and network variables are added to the set of exogenous covariates. Please see the set of references below for more
information of network (spatial) cross-sectional and panel models and their respective estimation strategies (e.g., fixed effects).

Moreover, the model can also estimate the network mixed effects model, where the estimated formula is:
\deqn{Y_t = \rho_1 W^{t}_{1}Y_t + \cdots + \rho_k W^{t}_{k}Y_t + X_t\beta + v_t, \ \ t = 1,2,\cdots,T }
\deqn{v_t = \mu_N + e_t,\ \ t = 1,2,\cdots,T   }
\deqn{e_t = \lambda W^e_t e_t + \epsilon_t, \ \ t = 1,2,\cdots,T }
where \eqn{\lambda} is the error autocorrelation parameter, \eqn{W^e_t} is the error social adjacency matrix (if applicable,
the network included in the \code{autocorrelated.network} argument), and \eqn{\epsilon_t} are a set of i.i.d normal errors at time \eqn{t}.

Internally, the function follows the two-step computation strategy discussed in Millo and Piras (2012). In the first step, the set of
parameter estimates for \eqn{\rho}, \eqn{\lambda}, and \eqn{\theta} are found via numerical optimization routines. Importantly, within each search,
the values for \eqn{\beta} and \eqn{\sigma^2} are updated via their GLS estimators. In the second step, once the ML parameter estimates for \eqn{\rho},
\eqn{\lambda}, and \eqn{\theta} are found, the final GLS estimates for \eqn{\beta} and \eqn{\sigma^2} are computed. The asymptotic
variance-covariance matrix for \eqn{\rho}, \eqn{\lambda}, and \eqn{\theta} are the inverse of the negative Hessian matrix and the
variance-covariance matrix for \eqn{\beta} is found via the standard GLS variance estimator. For sake of brevity, please see the below listed
references for more information on these estimators.
}
\examples{
data("net1", package = "netpanel") #the first panel network
data("net2", package = "netpanel") #the second panel network
data("net3", package = "netpanel") #the third panel network
data("simulated.data", package = "netpanel") #the simulated variables 

# a random effects panel network autocorrelation model
re.pnam <- mlpnam(Y~x1+x2+x3, net.formula = ~ net1,
               data =  simulated.data, time = ~panel,
               actor = ~unit, model = "random")
summary(re.pnam)              

# a fixed effects panel network autocorrelation model 
fe.pnam <- mlpnam(Y~x1+x2+x3, net.formula = ~  net1 + net2,
               data =  simulated.data, time = ~panel,
               actor = ~unit, model = "fixed",
               fixed.effect = "actor")
summary(fe.pnam)   

# a dynamic panel network autocorrelation model with autocorrelated errors
dyn.pnam <- mlpnam(Y~x1+x2+x3, net.formula = ~  net1 + net2,
               data =  simulated.data, time = ~panel,
               actor = ~unit, model = "dynamic",
               dynamic.lag = "outcome")
summary(dyn.pnam)   


\donttest{
# a fixed effects panel network autocorrelation model with autocorrelated errors
fe.pnam.error <- mlpnam(Y~x1+x2+x3, net.formula = ~ net1,
               data =  simulated.data, time = ~panel,
               actor = ~unit, model = "fixed",
               errors = "autocorrelated",
               autocorrelated.network = net3)
summary(fe.pnam.error)   
}


}
\references{
Anselin, Luc. 1988. \emph{Spatial Econometrics: Methods and Models}. Dordrecht: Springer Netherlands.

Cook, Scott J., Jude C. Hays and Robert J. Franzese. 2023. “STADL Up! The Spatiotemporal
Autoregressive Distributed Lag Model for TSCS Data Analysis.” \emph{American Political Science Review} 117(1):59–79.

Doreian, Patrick. 1982. “Maximum Likelihood Methods for Linear Models: Spatial Effect and
Spatial Disturbance Terms.” \emph{Sociological Methods & Research} 10(3):243–269.

Doriean, Patrick, Klaus Teuter and Chi-Hsein Wang. 1984. “Network Autocorrelation Models:
Some Monte Carlo Results.” \emph{Sociological Methods & Research} 13(2):155–200.

Duxbury, Scott. 2023. \emph{Longitudinal Network Models}. Vol. 192 of \emph{Quantitative Applications in the Social Sciences} SAGE Publications.

Hays, Jude C., Aya Kachi, and Robert J. Franzese. 2010. "A Spatial Model Incorporating
Dynamic, Endogenous Network Interdependence: A Political Science Application."
\emph{Statistical Methodology} 7(3): 406-428.

Lee, Lung-fei and Jihai Yu. 2010. “Estimation of Spatial Autoregressive Panel Data Models with
Fixed Effects.” \emph{Journal of Econometrics} 154(2):165–185.

Lee, Lung-fei and Jihai Yu. 2012a. “QML Estimation of Spatial Dynamic Panel Data Models with
Time Varying Spatial Weights Matrices.” \emph{Spatial Economic Analysis} 7(1):31–74.

Lee, Lung-fei and Jihai Yu. 2012b. “Spatial Panels: Random Components Versus Fixed Effects.”
\emph{International Economic Review} 53(4):1369–1412.

Millo, Giovanni. 2014. "Maximum Likelihood Estimation of Spatially and Serially Correlated
Panels with Random Effects." \emph{Computational Statistics & Data Analysis} 71: 914-933.

Millo, Giovanni and Gianfranco Piras. 2012. "splm: Spatial Panel Data Models in R."
\emph{Journal of Statistical Software} 47: 1-38.

Neuman, Eric J. and Mark S. Mizruchi. 2010. “Structure and Bias in the Network Autocorrelation
Model.” \emph{Social Networks} 32(4):290–300.

Wang, Wei and Jihai Yu. 2015. "Estimation of Spatial Panel Data Models with Time Varying
Spatial Weights Matrices.” \emph{Economic Letters} 128:95-99.
}
\seealso{
\code{\link{coef.pnam}}, \code{\link{vcov.pnam}}, \code{\link{logLik.pnam}},
\code{\link{AIC}}, \code{\link{residuals.pnam}}, \code{\link{fitted.pnam}},
\code{\link{netimpacts}}.
}
