% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pnam_wrapper_functions.R
\name{netimpacts}
\alias{netimpacts}
\title{The network impacts method (i.e., the matrix of impacts from the estimated model)}
\usage{
netimpacts(object, digits = 6)
}
\arguments{
\item{object}{An object of class "pnam".}

\item{...}{Additional arguments for other methods.}
}
\description{
This function returns the estimated network impacts from the
"pnam" fitted object.
}
