% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netpanel_wrapper_functions.R
\name{print.pnam}
\alias{print.pnam}
\title{Print Method for \code{pnam} model objects.}
\usage{
\method{print}{pnam}(x, digits = 5, ...)
}
\arguments{
\item{x}{An object of class \code{pnam}.}

\item{digits}{The number of digits to print after the decimal point.}

\item{...}{Additional arguments (currently unused).}
}
\value{
No return value. Prints out the main results of a \code{pnam} object.
}
\description{
Print Method for \code{pnam} model objects.
}
