% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pnam_wrapper_functions.R
\name{print.summary.pnam}
\alias{print.summary.pnam}
\title{Print Method for \code{summary.pnam} Panel Network Autocorrelation Model Summary}
\usage{
\method{print}{summary.pnam}(x, digits = 5, ...)
}
\arguments{
\item{x}{An object of class "summary.pnam".}

\item{digits}{The number of digits to print after the decimal point.}

\item{...}{Additional arguments (currently unused).}
}
\value{
No return value. Prints out the summary of a 'pnam' object.
}
\description{
Print Method for \code{summary.pnam} Panel Network Autocorrelation Model Summary
}
