% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pnam_wrapper_functions.R
\name{residuals.pnam}
\alias{residuals.pnam}
\title{Residuals for \code{pnam} model fits}
\usage{
\method{residuals}{pnam}(object, ...)
}
\arguments{
\item{object}{An object of class "pnam".}

\item{...}{Additional arguments for other methods.}
}
\description{
This function returns the model residuals for a \code{pnam} model fit.
}
