% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netpanel_data.R
\docType{data}
\name{simulated.data}
\alias{simulated.data}
\title{Simulated panel network dataset}
\format{
\subsection{\code{simulated.data}}{

A simulated panel network dataset with size \eqn{500 \times 5} the following variables: \eqn{Y}, \eqn{x_1},
\eqn{x_2}, \eqn{x_3}, \code{panel}, and \code{unit.}
}
}
\usage{
data(simulated.data)
}
\description{
A simulated panel network dataset that contains a response variable (\eqn{Y}) and three
exogenous regressors (\eqn{x_1}, \eqn{x_2}, and \eqn{x_3}). The exogenous regresors
\eqn{x_1} and \eqn{x_3} vectors are drawn from standard normal distribution, and
\eqn{x_2} is a vector of random bernoulli variables. The \code{panel}
variable represents the time of data collection and the \code{unit} variable indexes the
unique actors across the panels. The outcome is generated from the following deterministic
equation:
\deqn{Y_t = \rho_1 W^1 Y_t + \rho_2 W^w Y_t + \beta_0 + \beta_1 x_t1 + \beta_2 x_t2 + \beta_3 x_t3 + v_t, \ \ t = 1,2,\cdots,10}
\deqn{v_t = \mu_t + e_t, \ \ t = 1,2,\cdots,10}
\deqn{e_t = \lambda W^3 e_t + \epsilon_t, \ \ t = 1,2,\cdots,10}
where \eqn{W^1} is the \code{\link{net1}} \code{netpanel} data object, \eqn{W^2} is the  \code{\link{net2}} \code{netpanel} data object, and
\eqn{W^3} is the  \code{\link{net3}} \code{netpanel} data object.
}
\keyword{datasets}
