% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pnam_wrapper_functions.R
\name{vcov.pnam}
\alias{vcov.pnam}
\title{Extract variance-covariance matrix from \code{pnam} Fits}
\usage{
\method{vcov}{pnam}(object, ...)
}
\arguments{
\item{object}{An object of class "pnam".}

\item{...}{Additional arguments for other methods.}
}
\description{
This function extracts the variance-covariance matrix from estimated
\code{pnam} model fits for the exogenous covariates. For the endogenous
variance matrices (such as that for rho), please extract those
from the model fit. For example, the variance matrix for rho
can be extracted as \code{model$vcov.rho}.
}
\examples{
data("network.1", package = "netpanel") 
data("simulated.data", package = "netpanel")

# a fixed effects panel network autocorrelation model
fe.pnam <- mlpnam(Y~x1+x2+x3, net.formula = ~ net1,
               data =  simulated.data, time = ~panel,
               actor = ~unit, model = "fixed",
               fixed.effect = "actor")
vcov(fe.pnam)
}
