% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pnam_wrapper_functions.R
\name{vcov.pnam}
\alias{vcov.pnam}
\title{Extract variance-covariance matrix from `pnam` Fits}
\usage{
\method{vcov}{pnam}(object, ...)
}
\arguments{
\item{object}{An object of class "pnam".}

\item{...}{Additional arguments for other methods.}
}
\description{
This function extracts the variance-covariance matrix from estimated
`pnam` model fits for the exogenous covariates. For the endogenous 
variance matrices (such as that for rho), please extract those 
from the model fit. For example, the variance matrix for rho
can be extracted as `model$vcov.rho`.
}
